backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
httpx>=0.26.0
python-multipart>=0.0.7
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
playwright>=1.41.1
python-pptx>=0.6.23
faiss-cpu>=1.7.4
numpy>=1.26.3
jinja2>=3.1.3
requests>=2.31.0
websockets>=12.0
